requirements.txt
requests
slackclient=2.0.0
